requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
fonttools==4.45.1
frozenlist==1.4.0
googlesearch-python==1.2.3
idna==3.4
importlib-resources==6.1.1
install==1.3.5
kiwisolver==1.4.5
matplotlib==3.8.2
multidict==6.0.4
numpy==1.26.2
openai==0.28.0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
seaborn==0.13.0
six==1.16.0
sklearn==0.0.post11
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
yarl==1.9.2
zipp==3.17.0
